Add upgrading characters

+++ LegendaryScuffle/LegendaryScuffle/Character.h
@@ -4,8 +4,8 @@
 class Character
 {
 public:
-	float life;
-	float power;
+	double life;
+	double power;
 
 protected:
 	sf::Texture texture;
